fix(dev): Use requests.Session object for updating cookies
Before, we were not keeping track of cookie updates being sent by the
Piazza API when making requests. This is normally not an issue, but
presents a problem when a machine is making multiple requests to the
Piazza API for different networks. Imagine running a crawler for two
courses: when the second crawler begins crawling a different course, the
sessions for both crawlers become invalid because the server must be
sending updated cookies for them both in response to content.get calls.
By using a session object to make our requests, we can more simply
ensure that the cookies stay up to date, and this fixes the issues with
multiple crawlers on one machine.